cleanup, using hooks, no forced refresh

+++ src/Components/VisitorCard.js
@@ -1,31 +1,35 @@
-import React from 'react'
+import React, { useState, useEffect } from 'react'
 import { Link } from 'react-router-dom'
 import axios from 'axios'
-
-
 
 
 var votes = ["I am coming", "I am not coming", "I have not decided"];
 
 
-
-
 function VisitorCard(props) {
-    // console.log(props.visitor.name);
-    // props is redundant here
-    // here we are calling this component only to display its relevant data
-    // but we are not displaying a different url and the view remains in the
-    // same url as before. so we must pass the right object to this component
-    // so its data will be displayed. this is done by passing the object via props
-    function DeleteItem(e) {
-        e.preventDefault();
-        const url = `https://5f078d479c5c250016307111.mockapi.io/api/v1/visitors/${props.visitor.id}`
-        console.log("deleting" + url)
-        axios.delete(url)
-            .then(() => {window.location.reload(false)})
-            
-    }
 
+    const url = `https://5f078d479c5c250016307111.mockapi.io/api/v1/visitors/${props.visitor.id}`
+    // let request = { loading: false, data: null, error: false }
+
+    const [delete_visitor, setDeleteVisitor] = useState(false);
+
+    useEffect(() => {
+        // request = { loading: true, data: null, error: false }
+        console.log("VisitorCard::UseEffect called with " + delete_visitor)
+
+        if (delete_visitor) {
+            console.log("VisitorCard::UseEffect called with true")
+
+            setDeleteVisitor(false)
+            // console.log(`deleting visitor id: ${props.visitor.id}`)
+            axios.delete(url)
+                .then(response => {
+                    // request = { loading: false, data: response.data, error: false };
+                    props.setVisitorDeleted(true)
+                })
+                .catch(() => { /*request = { loading: false, data: null, error: true }*/ })
+        }
+    },[delete_visitor,props,url])
 
     function FunctionShowImage() {
         return (
@@ -84,7 +88,8 @@ function VisitorCard(props) {
     function FunctionDeleteButton() {
         return (
             <label className="bg-blue-500 text-white p-2 flex justify-center w-full"
-                onClick={DeleteItem}>Delete Me</label>
+                onClick={() => setDeleteVisitor(true)}>Delete Me</label>
+
         )
     }
